test av sending av epost til egen nav-epost
